requirements.txt
Flask-RESTful
flask-jwt-extended
Flask-SQLAlchemy
#uwsgi
psycopg2